mock Q3
Imagine you work for an online car dealership, and your company has recently received a large
shipment of cars of various makes and models. Your task is to create a C++ program to manage the
inventory and allow customers to easily find the cars they are interested in. You need to sort the cars
based on their models and prices.
Your program should allow users to perform the following operations:
1) Add Cars: Add new cars to the inventory.
2) List Cars by Model: Display a list of all cars sorted by model (alphabetically).
3) List Cars by Price: Display a list of all cars sorted by price (in ascending order).
4) List Cars by Engine size (CC): Display a list of all cars sorted by CC (in descending order).
5) Remove a Car: Remove a car from the inventory.